requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.81
langchain-text-splitters>=0.3.0
streamlit>=1.38.0
azure-search-documents>=11.4.0
python-dotenv>=1.0.0
pypdf>=4.0.0
